# specs/v1/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "nanocld",
    "description": "",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": ""
    },
    "version": "0.1.3"
  },
  "paths": {
    "/cargoes": {
      "get": {
        "tags": [
          "cargo"
        ],
        "summary": "List cargo",
        "description": "List cargo\n",
        "operationId": "list_cargo",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo are stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cargo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargoItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "cargo"
        ],
        "summary": "Create new cargo",
        "description": "Create new cargo\n",
        "operationId": "create_cargo",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo will be stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoPartial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "New cargo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/cargoes/count": {
      "get": {
        "tags": [
          "cargo"
        ],
        "summary": "Count cargo",
        "description": "Count cargo\n",
        "operationId": "count_cargo",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo is stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generic delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgGenericCount"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/cargoes/{name}": {
      "delete": {
        "tags": [
          "cargo"
        ],
        "summary": "Delete cargo by it's name",
        "description": "Delete cargo by it's name\n",
        "operationId": "delete_cargo_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the cargo",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo is stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generic delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgDeleteGeneric"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Api Error Structure that server send to client",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
      "CargoItem": {
        "type": "object",
        "description": "Cargo item is an definition to container create image and start them",
        "required": [
          "key",
          "namespace_name",
          "name",
          "image_name",
          "binds",
          "replicas"
        ],
        "properties": {
          "binds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dns_entry": {
            "type": "string"
          },
          "domainname": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "image_name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "namespace_name": {
            "type": "string"
          },
          "replicas": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CargoPartial": {
        "type": "object",
        "description": "Cargo partial",
        "required": [
          "name",
          "image_name"
        ],
        "properties": {
          "binds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dns_entry": {
            "type": "string"
          },
          "domainname": {
            "type": "string"
          },
          "environnements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hostname": {
            "type": "string"
          },
          "image_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "replicas": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}